Address #345 too many open files
Previous to this commit, all notebooklists queried the content
hierarchy (asynchronously) before allowing further interaction.
Depending on the size of the directory tree, this query can be
onerous.

This commit removes this requirement, and moves the query to
when it's needed.  It's not clear how widely used the content
hierarchy is (functions which need it include
`ein:notebooklist-find-server-by-notebook-name`
and `ein:notebooklist-open-notebook-global`).